Disable download mode and advise the user to run in local mode
Due to the source ODS data being removed from public download (it now requires registration and login to a portal), the download mode for the import tool doesn't currently work.
It is better to advise users of this when they try to use it and point them to an alternative method of running the import.